move events back to contract
