+++ exampleservice/CustomerService/Contract/CustomerRegisteredEvent.cs
@@ -1,7 +1,6 @@
-using exampleservice.CustomerService.Contract;
 using exampleservice.Framework.BaseFramework;
 
-namespace exampleservice.CustomerService.Events
+namespace exampleservice.CustomerService.Contract
 {
     public class CustomerRegisteredEvent : EventBase
     {
