+++ exampleservice/CustomerService/Contract/CustomerRegistrationFailedEvent.cs
@@ -1,7 +1,6 @@
-using exampleservice.CustomerService.Contract;
 using exampleservice.Framework.BaseFramework;
 
-namespace exampleservice.CustomerService.Events
+namespace exampleservice.CustomerService.Contract
 {
     public class CustomerRegistrationFailedEvent : EventBase
     {
