+++ exampleservice/CustomerService/Contract/InvalidSessionEvent.cs
@@ -1,7 +1,7 @@
 using System;
 using exampleservice.Framework.BaseFramework;
 
-namespace exampleservice.CustomerService.Events
+namespace exampleservice.CustomerService.Contract
 {
     public class InvalidSessionEvent : EventBase
     {
